refactor(engine-cdi): move static nested to dedicated classes
related to CAM-13735
