+++ engine-cdi/src/main/java/org/camunda/bpm/engine/cdi/impl/context/DefaultContextAssociationManager.java
@@ -57,106 +57,7 @@ import org.camunda.bpm.engine.variable.value.TypedValue;
 @SuppressWarnings("serial")
 public class DefaultContextAssociationManager implements ContextAssociationManager, Serializable {
 
-  private final static Logger log = Logger.getLogger(DefaultContextAssociationManager.class.getName());
-
-  protected static class ScopedAssociation {
-
-    @Inject
-    private RuntimeService runtimeService;
-
-    @Inject
-    private TaskService taskService;
-
-    protected VariableMap cachedVariables = new VariableMapImpl();
-    protected VariableMap cachedVariablesLocal = new VariableMapImpl();
-    protected Execution execution;
-    protected Task task;
-
-    public Execution getExecution() {
-      return execution;
-    }
-
-    public void setExecution(Execution execution) {
-      this.execution = execution;
-    }
-
-    public Task getTask() {
-      return task;
-    }
-
-    public void setTask(Task task) {
-      this.task = task;
-    }
-
-    @SuppressWarnings("unchecked")
-    public <T extends TypedValue> T getVariable(String variableName) {
-      TypedValue value = cachedVariables.getValueTyped(variableName);
-      if(value == null) {
-        if(execution != null) {
-          value = runtimeService.getVariableTyped(execution.getId(), variableName);
-          cachedVariables.put(variableName, value);
-        }
-      }
-      return (T) value;
-    }
-
-    public void setVariable(String variableName, Object value) {
-      cachedVariables.put(variableName, value);
-    }
-
-    public VariableMap getCachedVariables() {
-      return cachedVariables;
-    }
-
-    @SuppressWarnings("unchecked")
-    public <T extends TypedValue> T getVariableLocal(String variableName) {
-      TypedValue value = cachedVariablesLocal.getValueTyped(variableName);
-      if (value == null) {
-        if (task != null) {
-          value = taskService.getVariableLocalTyped(task.getId(), variableName);
-          cachedVariablesLocal.put(variableName, value);
-        } else if (execution != null) {
-          value = runtimeService.getVariableLocalTyped(execution.getId(), variableName);
-          cachedVariablesLocal.put(variableName, value);
-        }
-      }
-      return (T) value;
-    }
-
-    public void setVariableLocal(String variableName, Object value) {
-      if (execution == null && task == null) {
-        throw new ProcessEngineCdiException("Cannot set a local cached variable: neither a Task nor an Execution is associated.");
-      }
-      cachedVariablesLocal.put(variableName, value);
-    }
-
-    public VariableMap getCachedVariablesLocal() {
-      return cachedVariablesLocal;
-    }
-
-    public void flushVariableCache() {
-      if(task != null) {
-      	taskService.setVariablesLocal(task.getId(), cachedVariablesLocal);
-        taskService.setVariables(task.getId(), cachedVariables);
-
-      } else if(execution != null) {
-      	runtimeService.setVariablesLocal(execution.getId(), cachedVariablesLocal);
-        runtimeService.setVariables(execution.getId(), cachedVariables);
-
-      } else {
-        throw new ProcessEngineCdiException("Cannot flush variable cache: neither a Task nor an Execution is associated.");
-
-      }
-
-      // clear variable cache after flush
-      cachedVariables.clear();
-      cachedVariablesLocal.clear();
-    }
-
-  }
-
-  @ConversationScoped protected static class ConversationScopedAssociation extends ScopedAssociation implements Serializable {}
-  @RequestScoped protected static class RequestScopedAssociation extends ScopedAssociation implements Serializable {}
+  protected final static Logger log = Logger.getLogger(DefaultContextAssociationManager.class.getName());
 
   @Inject private BeanManager beanManager;
